+++ test-fixtures/env-setting-runner.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+CUSTOM_RUNNER_USED=1 exec node "$@"